@@ -443,7 +443,7 @@ namespace NEventStore
 
             protected CommitAttempt BuildCommitAttemptStub(int streamRevision, int commitSequence)
             {
-                List<EventMessage> events = new[] { new EventMessage() }.ToList();
+                EventMessage[] events = new[] { new EventMessage() };
                 return new CommitAttempt(Bucket.Default, _streamId, streamRevision, Guid.NewGuid(), commitSequence, SystemTime.UtcNow, null, events);
             }
 
