@@ -4,12 +4,12 @@
     using System.Collections.Generic;
     using System.Linq;
     using NEventStore.Persistence.AcceptanceTests.BDD;
-	using FluentAssertions;
+    using FluentAssertions;
 #if MSTEST
 	using Microsoft.VisualStudio.TestTools.UnitTesting;
 #endif
 #if NUNIT
-	using NUnit.Framework;	
+    using NUnit.Framework;
 #endif
 #if XUNIT
 	using Xunit;
@@ -19,7 +19,7 @@
 #if MSTEST
 	[TestClass]
 #endif
-	public class when_getting_from_to_then_should_not_get_later_commits : SpecificationBase
+    public class when_getting_from_to_then_should_not_get_later_commits : SpecificationBase
     {
         private readonly DateTime _endDate = new DateTime(2013, 1, 2);
         private readonly DateTime _startDate = new DateTime(2013, 1, 1);
@@ -31,8 +31,8 @@
             _engine = new InMemoryPersistenceEngine();
             _engine.Initialize();
             var streamId = Guid.NewGuid().ToString();
-            _engine.Commit(new CommitAttempt(streamId, 1, Guid.NewGuid(), 1, _startDate, new Dictionary<string, object>(), new List<EventMessage>{ new EventMessage()}));
-            _engine.Commit(new CommitAttempt(streamId, 2, Guid.NewGuid(), 2, _endDate, new Dictionary<string, object>(), new List<EventMessage>{ new EventMessage()}));
+            _engine.Commit(new CommitAttempt(streamId, 1, Guid.NewGuid(), 1, _startDate, new Dictionary<string, object>(), new EventMessage[] { new EventMessage() }));
+            _engine.Commit(new CommitAttempt(streamId, 2, Guid.NewGuid(), 2, _endDate, new Dictionary<string, object>(), new EventMessage[] { new EventMessage() }));
         }
 
         protected override void Because()
