@@ -1,14 +1,10 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading;
 
 namespace NEventStore.Helpers
 {
     internal sealed class DisposableAction : IDisposable
     {
-
         public static readonly DisposableAction Empty = new DisposableAction(null);
 
         private Action _disposeAction;
@@ -22,11 +18,7 @@ namespace NEventStore.Helpers
         {
             // Interlocked allows the continuation to be executed only once
             Action dispose = Interlocked.Exchange(ref _disposeAction, null);
-            if (dispose != null)
-            {
-                dispose();
-            }
+            dispose?.Invoke();
         }
-
     }
 }
