@@ -1,21 +1,12 @@
+#pragma warning disable RCS1170 // Use read-only auto-implemented property.
+
 namespace NEventStore.Persistence
 {
     using System;
     using System.Collections.Generic;
-    using System.Collections.ObjectModel;
 
     public class Commit : ICommit
     {
-        private readonly string _bucketId;
-        private readonly string _streamId;
-        private readonly int _streamRevision;
-        private readonly Guid _commitId;
-        private readonly int _commitSequence;
-        private readonly DateTime _commitStamp;
-        private readonly IDictionary<string, object> _headers;
-        private readonly ICollection<EventMessage> _events;
-        private readonly Int64 _checkpointToken;
-
         public Commit(
             string bucketId,
             string streamId,
@@ -25,70 +16,40 @@ namespace NEventStore.Persistence
             DateTime commitStamp,
             Int64 checkpointToken,
             IDictionary<string, object> headers,
-            IEnumerable<EventMessage> events)
+            ICollection<EventMessage> events)
         {
-            _bucketId = bucketId;
-            _streamId = streamId;
-            _streamRevision = streamRevision;
-            _commitId = commitId;
-            _commitSequence = commitSequence;
-            _commitStamp = commitStamp;
-            _checkpointToken = checkpointToken;
-            _headers = headers ?? new Dictionary<string, object>();
-            _events = events == null ?
-                new ReadOnlyCollection<EventMessage>(new List<EventMessage>()) :
-                new ReadOnlyCollection<EventMessage>(new List<EventMessage>(events));
+            BucketId = bucketId;
+            StreamId = streamId;
+            StreamRevision = streamRevision;
+            CommitId = commitId;
+            CommitSequence = commitSequence;
+            CommitStamp = commitStamp;
+            CheckpointToken = checkpointToken;
+            Headers = headers ?? new Dictionary<string, object>();
+            Events = events ?? Array.Empty<EventMessage>();
+            //Events = events == null ?
+            //    new ReadOnlyCollection<EventMessage>(new List<EventMessage>()) :
+            //    new ReadOnlyCollection<EventMessage>(new List<EventMessage>(events));
         }
 
-        public string BucketId
-        {
-            get { return _bucketId; }
-        }
+        public string BucketId { get; private set; }
 
-        public string StreamId
-        {
-            get { return _streamId; }
-        }
+        public string StreamId { get; private set; }
 
-        public int StreamRevision
-        {
-            get { return _streamRevision; }
-        }
+        public int StreamRevision { get; private set; }
 
-        public Guid CommitId
-        {
-            get { return _commitId; }
-        }
+        public Guid CommitId { get; private set; }
 
-        public int CommitSequence
-        {
-            get { return _commitSequence; }
-        }
+        public int CommitSequence { get; private set; }
 
-        public DateTime CommitStamp
-        {
-            get { return _commitStamp; }
-        }
+        public DateTime CommitStamp { get; private set; }
 
-        public IDictionary<string, object> Headers
-        {
-            get { return _headers; }
-        }
+        public IDictionary<string, object> Headers { get; private set; }
 
-        public ICollection<EventMessage> Events
-        {
-            get
-            {
-                return _events;
-            }
-        }
+        public ICollection<EventMessage> Events { get; private set; }
 
-        public Int64 CheckpointToken
-        {
-            get
-            {
-                return _checkpointToken;
-            }
-        }
+        public Int64 CheckpointToken { get; private set; }
     }
 }
+
+#pragma warning restore RCS1170 // Use read-only auto-implemented property.
