@@ -92,14 +92,14 @@ namespace NEventStore.Persistence
 
         /// <summary>
         ///     Completely DESTROYS the contents of ANY and ALL streams that have been successfully persisted
-        ///     in the specified bucket. 
+        ///     in the specified bucket.
         ///     Use with caution.
         /// </summary>
         void Purge(string bucketId);
 
         /// <summary>
         ///     Completely DESTROYS the contents and schema (if applicable) containting ANY and ALL streams that have been
-        ///     successfully persisted. 
+        ///     successfully persisted.
         ///     Use with caution.
         /// </summary>
         void Drop();
